# specs/001-auth-registration-onboarding/contracts/api.yaml
openapi: 3.0.3
info:
  title: Plinto Auth and Onboarding API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - CookieAuth: []
paths:
  /auth/session:
    post:
      summary: Create application session from OIDC identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      summary: Revoke the current session
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: plinto_session
  responses:
...
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Authentication required
        details:
          type: array
          items:
            type: object
        traceId:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error"
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]
...
    Membership:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [owner, member, viewer]
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, userId, role, createdAt]
...
    CreateSessionRequest:
      type: object
      properties:
        idpSub:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
      required: [idpSub, email]
...
    CreateSessionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            sessionId:
              type: string
            expiresAt:
              type: string
              format: date-time
            user:
              $ref: "#/components/schemas/User"
            memberships:
              type: array
              items:
                $ref: "#/components/schemas/Membership"
            activeTenantId:
              type: string
              nullable: true
            needsOnboarding:
              type: boolean
          required: [sessionId, expiresAt, user, memberships, needsOnboarding]
      required: [data]
    LogoutResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            success:
              type: boolean
          required: [success]
      required: [data]